+++ R/NMreadCsv.R
@@ -13,21 +13,16 @@
 ##'     something else. If data.tables are wanted, use
 ##'     as.fun="data.table". The default can be configured using
 ##'     NMdataConf.
-##' @param format Format of file to read. Must be one of "csv", "rds",
-##'     or "fst". Default is to determine this from the file name
-##'     extension. Notice, if a delimited format is used, the
-##'     extension can very well be different from "csv" (say file name
-##'     is "input.tab"), and one must provide format="csv". This will
-##'     work for any delimited format supported by fread and does not
-##'     need to be comma-separated even though the format name
-##'     indicates so.
-##' @param args.fst Optional arguments to pass to \code{read_fst} if
-##'     code{format="fst"} is used.
-##' @details This is almost just a shortcut to fread so you don't have
-##'     to remember how to read the data that was exported for
-##'     Nonmem. The only added feature is that meta data as written by
-##'     NMwriteData is read and attached as NMdata metadata before
-##'     data is returned.
+##' @param format Format of file to read. Can be of length>1 in which
+##'     case the first format found will be used (i.e. format is a
+##'     prioritized vector). If not one of "rds" or "fst", it is
+##'     assumed to be a delimited text file. Default is to determine
+##'     this from the file name extension. Notice, if a delimited
+##'     format is used, the extension can very well be different from
+##'     "csv" (say file name is
+##'     "input.tab")". This will work for any delimited format supported by fread.
+##' @param args.fst Optional arguments to pass to \code{read_fst} if code{format="fst"} is used.
+##' @details This is almost just a shortcut to fread so you don't have to remember how to read the data that was exported for Nonmem. The only added feature is that meta data as written by NMwriteData is read and attached as NMdata metadata before data is returned.
 ##' @return A data set of class as defined by as.fun.
 ##' @importFrom data.table fread
 ##' @importFrom fst read_fst
@@ -38,12 +33,30 @@
 NMreadCsv <- function(file,args.fread,as.fun=NULL,format=fnExtension(file),args.fst){
 
     
-    if( !is.character(format) && length(format) == 1 ){
-        stop("format must be a single character string.")
+    if( !is.character(format) ){
+        stop("format must be character strings.")
     }
-    if(!format%in%c("rds","fst")) format <- "text"
-    
     as.fun <- NMdataDecideOption("as.fun",as.fun)
+
+### this requires format to be only one of text, rds or fst
+    ## if( !is.character(format) && length(format) == 1 ){
+    ##     stop("format must be a single character string.")
+    ## }
+    ## if(!format%in%c("rds","fst")) format <- "text"
+
+### we take the first file available. format is file name
+### extension. If not fst or rds, it is expected to be a delimited
+### text file.
+    format <- unique(format)
+    exts <- format
+    files <- fnExtension(file,exts)
+    exists <- file.exists(files)
+    exts <- exts[exists][1]
+    format[!format%in%c("rds","fst")] <- "text"
+    format <- format[exists][1]
+    file <- files[exists][1]
+
+
 
     dt <- switch(format,
                  text={